pyproject.toml
[tool.poetry]
name = "moschitta-query"
version = "0.1.0"
description = "Fluent Query Builder Library"
authors = ["Skyler Saville <skylersaville@gmail.com>"]
readme = "README.md"
packages = [{include = "moschitta_query"}]

[tool.poetry.dependencies]
python = "^3.11"
psycopg2-binary = "^2.9.9"
mysql-connector-python = "^8.4.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
isort = "^5.13.2"
pylint = "^3.1.0"
pytest = "^8.2.0"
mypy = "^1.10.0"
sphinx = "^7.3.7"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
